git-rebaset: git rebase while making sure that no two timestamp match
GitHub shows commits in commit timestamp order and not in commit order.
Because rebasing can be fast and the timestamp is only precise
up to seconds, many, if not all commits end up having the same timestamp.
Hence a pull request rebased to master becomes a mess to review.

If this script sees a commit that would have the same timestamp as the
previous one, it increments it by one second.

Work-in-progress quality, but works for me.

Signed-off-by: bkil-syslogng <tamas.nagy@balabit.com>